bashrc: fix $COLUMNS & $LINES after each child exit.
When the user change the window size of terminal,
bash should receive a SIGWIDCH signal and
bash will change $COLUMNS and $LINES to currect value.

But if there are a subprocess running, bash will not
receive the SIGWIDCH, the subprocess will receive it.
So bash will not change $COLUMNS and $LINES.
It seems like a BUG of X Window System.

In this patch, we force the bash fix $COLUMNS & $LINES
to currect value after each child process exited (receive
SIGCHLD signal).